Maintenance Task
* Split SymbolId.h into SymbolId.h & NodeId.h
* Removed all unused functions from both SymbolId & NodeId classes
* NodeId is meant to be an internal class i.e. not to be published
  during installation.
