+++ include/Surelog/API/SLAPI.h
@@ -25,7 +25,7 @@
 #define SURELOG_SLAPI_H
 #pragma once
 
-#include <Surelog/Common/SymbolId.h>
+#include <Surelog/Common/NodeId.h>
 #include <Surelog/SourceCompile/VObjectTypes.h>
 
 #include <string>
