+++ src/DesignCompile/CompileToolbox.cpp
@@ -21,6 +21,7 @@
  * Created on June 13, 2018, 9:36 PM
  */
 
+#include <Surelog/Common/NodeId.h>
 #include <Surelog/DesignCompile/CompileToolbox.h>
 
 namespace SURELOG {
